added pyproject, updated README.md, refactored project structure
