+++ pyproject.toml
@@ -1,0 +1,56 @@
+[build-system]
+requires      = ["setuptools>=61.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "wizards-remembrall"
+version = "1.0.0"
+description = "Remembrall is an assistant bot for wizards of the academy"
+readme = "README.md"
+authors = [
+    { name = "GoIT :: MCS CyberSecurity 2 :: Project Group 2", email = "info@goit.ua" },
+]
+license = { file = "LICENSE" }
+classifiers = [
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+]
+keywords = ["notes", "addressbook", "contacts", "assistant", "cli"]
+dependencies = []
+requires-python = ">=3.11"
+
+    [project.optional-dependencies]
+    build = ["build", "twine"]
+    dev   = ["black", "bumpver", "isort", "mypy", "pytest"]
+
+    [project.scripts]
+    remembrall = "remembrall.__main__:main"
+
+    [project.urls]
+    repository    = "https://github.com/ihahalev/wizard-assistant-bot"
+
+[tool.bumpver]
+current_version = "1.0.0"
+version_pattern = "MAJOR.MINOR.PATCH"
+commit_message  = "bump version {old_version} -> {new_version}"
+commit          = true
+tag             = true
+push            = false
+
+    [tool.bumpver.file_patterns]
+    "pyproject.toml" = [
+        'current_version = "{version}"',
+        'version = "{version}"',
+    ]
+    "src/remembrall/__init__.py" = ["{version}"]
+    "src/remembrall/__main__.py" = ["- wizards-remembrall v{version}"]
+
+[tool.isort]
+profile                   = "black"
+import_heading_stdlib     = "Standard library imports"
+import_heading_thirdparty = "Third party imports"
+import_heading_firstparty = "Reader imports"
+
+[tool.mypy]
+strict = true
