+++ src/remembrall/__init__.py
@@ -1,0 +1,10 @@
+"""Assistant bot for wizards of the academy.
+"""
+
+# Standard library imports
+import sys
+
+sys.path.append('src/remembrall')
+
+# Version of wizards-remembrall package
+__version__ = "1.0.0"
