+++ src/remembrall/__main__.py
@@ -1,6 +1,22 @@
-from classes import Record
-from helpers import book_operations
-from helpers.data_upload import load_data, save_data
+"""Assistant bot for wizards of the academy.
+
+Usage:
+--------
+
+    $ remembrall [command] [args1, args2, ...]
+
+List the avaialble commands:
+
+    $ remembrall help
+
+Version:
+--------
+
+- wizards-remembrall v1.0.0
+"""
+from .classes import Record
+from .helpers import book_operations
+from .helpers.data_upload import load_data, save_data
 
 def parse_input(user_input: str) -> tuple:
     cmd, *args = user_input.split()
