+++ src/remembrall/classes/Record.py
