+++ src/remembrall/classes/customErrors.py
