+++ src/remembrall/helpers/book_operations.py
